+++ paper.bib
@@ -17,8 +17,13 @@
         the {Koopman} Operator},
     author = {Pan, Shaowu and Kaiser, Eurika and de Silva, Brian M. and Kutz, J.
               Nathan and Brunton, Steven L.},
-    year = {2023},
-    journaltitle = {{\tt arXiv:2306.12962v1[eess.SY]}},
+    journal = {J. Open Source Softw.},
+    year = {2024},
+    doi = {10.21105/joss.05881},
+    publisher = {The Open Journal},
+    volume = {9},
+    number = {94},
+    pages = {5881},
 }
 
 @inproceedings{dlkoopman,
@@ -33,7 +38,7 @@
 
 @misc{kooplearn,
     author = {Pietro Novelli and Vladimir Kostic},
-    title = {\texttt{{Machine-Learning-Dynamical-Systems/kooplearn}} v1.1.0},
+    title = {{Machine-Learning-Dynamical-Systems/kooplearn}} v1.1.0},
     year = 2023,
     url = {https://github.com/Machine-Learning-Dynamical-Systems/kooplearn},
     urldate = {2024-08-06},
@@ -69,6 +74,7 @@
     author = {Lortie, Louis and Dahdah, Steven and Forbes, James Richard},
     year = {2024},
     journaltitle = {{\tt arXiv:2403.10623v1[eess.SY]}},
+    doi = {10.48550/arXiv.2403.10623},
 }
 
 @article{dahdah_2024_uncertainty,
@@ -76,6 +82,7 @@
     author = {Dahdah, Steven and Forbes, James Richard},
     year = {2024},
     journaltitle = {{\tt arXiv:2410.01057v1[eess.SY]}},
+    doi = {10.48550/arXiv.2410.01057},
 }
 
 @article{lortie_2024_asymptotically,
@@ -84,6 +91,7 @@
     author = {Lortie, Louis and Forbes, James Richard},
     year = {2024},
     journaltitle = {{\tt arXiv:2408.16846v1[eess.SY]}},
+    doi = {10.48550/arXiv.2408.16846},
 }
 
 @misc{pykoop,
