+++ paper.bib
@@ -38,7 +38,7 @@
 
 @misc{kooplearn,
     author = {Pietro Novelli and Vladimir Kostic},
-    title = {{Machine-Learning-Dynamical-Systems/kooplearn}} v1.1.0},
+    title = {{Machine-Learning-Dynamical-Systems/kooplearn} v1.1.0},
     year = 2023,
     url = {https://github.com/Machine-Learning-Dynamical-Systems/kooplearn},
     urldate = {2024-08-06},
