### 2024-11-12 #7451 @sdahdah
The remaining references do not have DOIs as far as I can tell.
