### 2025-03-26 #7947 @editorialbot
[REVIEW]: pykoop: a Python Library for Koopman Operator Approximation
**Submitting author:** @sdahdah (<a href="http://orcid.org/0000-0003-4930-9634">Steven Dahdah</a>)
**Repository:** https://github.com/decargroup/pykoop
**Branch with paper.md** (empty if default branch): paper
**Version:** v2.0.1
