### 2025-05-09 decargroup/pykoop#187 @matt-graham
Minor comments on JOSS paper text
Raising as part of openjournals/joss-reviews#7947

The paper in general reads very well. I have a few very minor comments:

- As the repository is already linked to (under _Software_) in the 'sidebar' of the rendered JOSS paper, having a link to the repository in the summary as well feels a bit redundant. I'm not sure if there is a JOSS convention on this, and I am currently reviewing another paper which does same as here by linking to repository in _Summary_ so it may be well be that I'm in the wrong in finding this odd. Either way feel free to ignore this as it is just a nitpick!
- In the section _Scholarly publications using `pykoop`_, as the citations are being directly references in the text I think it would be better to use the non-parenthesized form ([achieved my omitting the square brackets in the Markdown source](https://pandoc.org/MANUAL.html#extension-citations)).
- I believe from the note at the end of the [example paper in the JOSS documentation](https://joss.readthedocs.io/en/latest/example_paper.html#example-paper)

  > Also note that the references include full names of venues, e.g., journals and conferences, not abbreviations only understood in the context of a specific discipline.

  that the expectation is that all journal names in references should be written in full rather than abbreviated. Currently the paper uses various abbreviations - for example _Proc. R. Soc. A._ &rightarrow; _Proceedings of Royal Society A_ and _J. Open Source Softw._ &rightarrow; _Journal of Open Source Software_
