### 2025-05-09 decargroup/pykoop#188 @matt-graham
Some tests failing when running locally in Python 3.13 environment
Raising as part of openjournals/joss-reviews#7947

# Expected Behavior

All tests pass.

# Actual Behavior

Some (64) tests failed.

<summary>pytest short test summary info output</summary>

```
=================================================================================================================== short test summary info ====================================================================================================================
FAILED tests/test_kernel_approximation.py::TestSkLearn::test_compatible_estimator[RandomFourierKernelApprox(random_state=1234)-check_n_features_in_after_fitting] - AssertionError: `RandomFourierKernelApprox.transform()` does not check for consistency between input number
FAILED tests/test_kernel_approximation.py::TestSkLearn::test_compatible_estimator[RandomFourierKernelApprox(method='weight_only',random_state=1234)-check_n_features_in_after_fitting] - AssertionError: `RandomFourierKernelApprox.transform()` does not check for consistency between input number
FAILED tests/test_kernel_approximation.py::TestSkLearn::test_compatible_estimator[RandomBinningKernelApprox(random_state=1234)-check_n_features_in_after_fitting] - AssertionError: `RandomBinningKernelApprox.transform()` does not check for consistency between input number
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[KoopmanPipeline(regressor=Edmd())-check_do_not_raise_errors_in_init_or_set_params] - ValueError: not enough values to unpack (expected 2, got 1)
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[KoopmanPipeline(regressor=Edmd())-check_n_features_in_after_fitting] - AssertionError: `KoopmanPipeline.predict()` does not check for consistency between input number
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[KoopmanPipeline(lifting_functions=[('pl',PolynomialLiftingFn())],regressor=Edmd())-check_do_not_raise_errors_in_init_or_set_params] - ValueError: not enough values to unpack (expected 2, got 1)
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[KoopmanPipeline(lifting_functions=[('pl',PolynomialLiftingFn())],regressor=Edmd())-check_n_features_in_after_fitting] - AssertionError: `KoopmanPipeline.predict()` does not check for consistency between input number
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[SplitPipeline()-check_do_not_raise_errors_in_init_or_set_params] - ValueError: not enough values to unpack (expected 2, got 1)
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[SplitPipeline()-check_n_features_in_after_fitting] - AssertionError: `SplitPipeline.transform()` does not check for consistency between input number
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[SplitPipeline(lifting_functions_state=[('pl',PolynomialLiftingFn())])-check_do_not_raise_errors_in_init_or_set_params] - ValueError: not enough values to unpack (expected 2, got 1)
FAILED tests/test_koopman_pipeline.py::TestSkLearn::test_compatible_estimator[SplitPipeline(lifting_functions_state=[('pl',PolynomialLiftingFn())])-check_n_features_in_after_fitting] - AssertionError: `SplitPipeline.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[PolynomialLiftingFn()-check_n_features_in_after_fitting] - AssertionError: `PolynomialLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[DelayLiftingFn()-check_n_features_in_after_fitting] - AssertionError: `DelayLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[SkLearnLiftingFn(transformer=MaxAbsScaler())-check_n_features_in_after_fitting] - AssertionError: `SkLearnLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[BilinearInputLiftingFn()-check_n_features_in_after_fitting] - AssertionError: `BilinearInputLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[RbfLiftingFn(centers=QmcCenters(random_state=1234))-check_n_features_in_after_fitting] - AssertionError: `RbfLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[ConstantLiftingFn()-check_n_features_in_after_fitting] - AssertionError: `ConstantLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_lifting_functions.py::TestSkLearn::test_compatible_estimator[KernelApproxLiftingFn(kernel_approx=RandomFourierKernelApprox(random_state=1234))-check_n_features_in_after_fitting] - AssertionError: `KernelApproxLiftingFn.transform()` does not check for consistency between input number
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Edmd()-check_estimator_tags_renamed] - TypeError: Estimator Edmd has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` or `_get_t...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Edmd()-check_n_features_in_after_fitting] - AssertionError: `Edmd.predict()` does not check for consistency between input number
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta()-check_estimator_tags_renamed] - TypeError: Estimator EdmdMeta has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` or `_g...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta()-check_n_features_in_after_fitting] - AssertionError: `EdmdMeta.predict()` does not check for consistency between input number
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_estimator_tags_renamed] - TypeError: Estimator EdmdMeta has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` or `_g...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_fit_score_takes_y] - ValueError: Found input variables with inconsistent numbers of samples: [30, 1]
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_n_features_in_after_fitting] - AssertionError: `EdmdMeta.predict()` does not check for consistency between input number
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_pipeline_consistency] - ValueError: Found input variables with inconsistent numbers of samples: [30, 1]
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_regressors_train] - AssertionError
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_regressors_train(readonly_memmap=True)] - AssertionError
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_regressors_train(readonly_memmap=True,X_dtype=float32)] - AssertionError
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[EdmdMeta(regressor=Ridge(alpha=1))-check_methods_sample_order_invariance] - IndexError: index 5 is out of bounds for axis 0 with size 1
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmdc()-check_estimator_tags_renamed] - TypeError: Estimator Dmdc has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` or `_get_t...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmdc()-check_n_features_in_after_fitting] - AssertionError: `Dmdc.predict()` does not check for consistency between input number
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimator_tags_renamed] - TypeError: Estimator Dmd has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` or `_get_ta...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_fit_score_takes_y] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_overwrite_params] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_dont_overwrite_parameters] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_fit_returns_self] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_readonly_memmap_input] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_n_features_in_after_fitting] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 4)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_positive_only_tag_during_fit] - AssertionError: Estimator 'Dmd' raised ValueError unexpectedly. This happens when passing negative input values as X. If negative values are not supported for this estimator instance, then the tags.input_tags.positive_only tag needs to be set to True.
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_dtypes] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 5)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_dtype_object] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_pipeline_consistency] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_nan_inf] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_pickle] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_estimators_pickle(readonly_memmap=True)] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_f_contiguous_array_estimator] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressors_train] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressors_train(readonly_memmap=True)] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressors_train(readonly_memmap=True,X_dtype=float32)] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressor_data_not_an_array] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressor_multioutput] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 5 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressors_no_decision_function] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 4)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_regressors_int] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 10)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_methods_sample_order_invariance] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_methods_subset_invariance] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_dict_unchanged] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_fit_idempotent] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_fit_check_is_fitted] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_n_features_in] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 2)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[Dmd()-check_fit2d_predict1d] - ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 1 is different from 3)
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[DataRegressor()-check_estimator_tags_renamed] - TypeError: Estimator DataRegressor has defined either `_more_tags` or `_get_tags`, but not `__sklearn_tags__`. If you're customizing tags, and need to support multiple scikit-learn versions, you can implement both `__sklearn_tags__` and `_more_tags` o...
FAILED tests/test_regressors.py::TestSkLearn::test_compatible_estimator[DataRegressor()-check_n_features_in_after_fitting] - AssertionError: `DataRegressor.predict()` does not check for consistency between input number
FAILED tests/test_util.py::TestSkLearn::test_compatible_estimator[AnglePreprocessor()-check_n_features_in_after_fitting] - AssertionError: `AnglePreprocessor.transform()` does not check for consistency between input number
========================================================================================== 64 failed, 1819 passed, 474 deselected, 1889 warnings in 412.30s (0:06:52) ==========================================================================================
```

# Steps to Reproduce the Problem

From a clone of repository at commit 2215358

1. Create a blank Python 3.13 virtual environment - I used `uv` - and activate
   ```
   uv venv --python=3.13
   source .venv/bin/activate
   ```
2. Install development dependencies and local package
   ```
   uv pip install -r requirements-dev.txt
   uv pip install .
   ```
3. Run tests (skipping MOSEK dependent tests)
   ```
   pytest tests -k "not mosek"
   ```

# Specifications

- Package version: Current `main` ([2215358](https://github.com/decargroup/pykoop/commit/22153581483c1d36afc4226e45a1efa42d464bf2))
- Python version: 3.13
- Platform: Ubuntu 24.04
