### 2025-05-09 decargroup/pykoop#188 @matt-graham
As tests appear to have all passed in [the last GitHub Actions workflow run](https://github.com/decargroup/pykoop/actions/runs/11941124138) across Python versions 3.8 to 3.11, with this run being 6 months ago. It's not clear therefore if the tests are failing due to something changing in an (unpinned) upstream dependency, a Python 3.13 specific issue or something else.
