### 2025-05-09 decargroup/pykoop#188 @matt-graham
Looks like this is not specific to Python 3.13, as I also get failures when running in a Python 3.11 environment:

```
uv venv --python=3.11
source .venv/bin/activate
uv pip install -r requirements-dev.txt
uv pip install .
pytest tests -k "not mosek" -x
```

fails and exits with

```
FAILED tests/test_kernel_approximation.py::TestSkLearn::test_compatible_estimator[RandomFourierKernelApprox(random_state=1234)-check_n_features_in_after_fitting] - AssertionError: `RandomFourierKernelApprox.transform()` does not check for consistency between input number
```

There are also a series of deprecation warnings raised:

<summary>Full pytest error traceback info</summary>

```
____________________________________________________________________ TestSkLearn.test_compatible_estimator[RandomFourierKernelApprox(random_state=1234)-check_n_features_in_after_fitting] _____________________________________________________________________

name = 'RandomFourierKernelApprox', estimator_orig = RandomFourierKernelApprox(random_state=1234)

    @ignore_warnings(category=FutureWarning)
    def check_n_features_in_after_fitting(name, estimator_orig):
        # Make sure that n_features_in are checked after fitting
        tags = get_tags(estimator_orig)

        is_supported_X_types = tags.input_tags.two_d_array or tags.input_tags.categorical

        if not is_supported_X_types or tags.no_validation:
            return

        rng = np.random.RandomState(0)

        estimator = clone(estimator_orig)
        set_random_state(estimator)
        if "warm_start" in estimator.get_params():
            estimator.set_params(warm_start=False)

        n_samples = 10
        X = rng.normal(size=(n_samples, 4))
        X = _enforce_estimator_tags_X(estimator, X)

        if is_regressor(estimator):
            y = rng.normal(size=n_samples)
        else:
            y = rng.randint(low=0, high=2, size=n_samples)
        y = _enforce_estimator_tags_y(estimator, y)

        err_msg = (
            "`{name}.fit()` does not set the `n_features_in_` attribute. "
            "You might want to use `sklearn.utils.validation.validate_data` instead "
            "of `check_array` in `{name}.fit()` which takes care of setting the "
            "attribute.".format(name=name)
        )

        estimator.fit(X, y)
        assert hasattr(estimator, "n_features_in_"), err_msg
        assert estimator.n_features_in_ == X.shape[1], err_msg

        # check methods will check n_features_in_
        check_methods = [
            "predict",
            "transform",
            "decision_function",
            "predict_proba",
            "score",
        ]
        X_bad = X[:, [1]]

        err_msg = """\
            `{name}.{method}()` does not check for consistency between input number
            of features with {name}.fit(), via the `n_features_in_` attribute.
            You might want to use `sklearn.utils.validation.validate_data` instead
            of `check_array` in `{name}.fit()` and {name}.{method}()`. This can be done
            like the following:
            from sklearn.utils.validation import validate_data
            ...
            class MyEstimator(BaseEstimator):
                ...
                def fit(self, X, y):
                    X, y = validate_data(self, X, y, ...)
                    ...
                    return self
                ...
                def {method}(self, X):
                    X = validate_data(self, X, ..., reset=False)
                    ...
                return X
        """
        err_msg = textwrap.dedent(err_msg)

        msg = f"X has 1 features, but \\w+ is expecting {X.shape[1]} features as input"
        for method in check_methods:
            if not hasattr(estimator, method):
                continue

            callable_method = getattr(estimator, method)
            if method == "score":
                callable_method = partial(callable_method, y=y)

            with raises(
                ValueError, match=msg, err_msg=err_msg.format(name=name, method=method)
            ):
>               callable_method(X_bad)

.venv/lib/python3.11/site-packages/sklearn/utils/estimator_checks.py:4501:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
.venv/lib/python3.11/site-packages/sklearn/utils/_set_output.py:319: in wrapped
    data_to_wrap = f(self, X, *args, **kwargs)
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = RandomFourierKernelApprox(random_state=0)
X = array([[ 0.40015721],
       [-0.97727788],
       [ 0.4105985 ],
       [ 0.12167502],
       [-0.20515826],
       [ 0.6536186 ],
       [-1.45436567],
       [ 1.46935877],
       [-1.98079647],
       [ 1.20237985]])

    def transform(self, X: np.ndarray) -> np.ndarray:
        """Transform data.

        Parameters
        ----------
        X : np.ndarray
            Data matrix.

        Returns
        -------
        np.ndarray
            Transformed data matrix.
        """
        sklearn.utils.validation.check_is_fitted(self)
        X = sklearn.utils.validation.check_array(X)
        X_scaled = np.sqrt(2 * self.shape) * X
>       products = X_scaled @ self.random_weights_  # (n_samples, n_components)
E       ValueError: matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 4 is different from 1)

pykoop/kernel_approximation.py:271: ValueError

self = <test_kernel_approximation.TestSkLearn object at 0x7b7a3a8e7210>, estimator = RandomFourierKernelApprox(random_state=1234), check = functools.partial(<function check_n_features_in_after_fitting at 0x7b7a3ab4f060>, 'RandomFourierKernelApprox')

    @sklearn.utils.estimator_checks.parametrize_with_checks([
        pykoop.RandomFourierKernelApprox(
            method='weight_offset',
            random_state=1234,
        ),
        pykoop.RandomFourierKernelApprox(
            method='weight_only',
            random_state=1234,
        ),
        pykoop.RandomBinningKernelApprox(random_state=1234),
    ])
    def test_compatible_estimator(self, estimator, check):
        """Test ``scikit-learn`` compatibility of estimators."""
>       check(estimator)

tests/test_kernel_approximation.py:435:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
.venv/lib/python3.11/site-packages/sklearn/utils/_testing.py:147: in wrapper
    return fn(*args, **kwargs)
.venv/lib/python3.11/site-packages/sklearn/utils/estimator_checks.py:4498: in check_n_features_in_after_fitting
    with raises(
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <sklearn.utils._testing._Raises object at 0x7b7a204dbb50>, exc_type = <class 'ValueError'>
exc_value = ValueError('matmul: Input operand 1 has a mismatch in its core dimension 0, with gufunc signature (n?,k),(k,m?)->(n?,m?) (size 4 is different from 1)'), _ = <traceback object at 0x7b7a202080c0>

    def __exit__(self, exc_type, exc_value, _):
        # see
        # https://docs.python.org/2.5/whatsnew/pep-343.html#SECTION000910000000000000000

        if exc_type is None:  # No exception was raised in the block
            if self.may_pass:
                return True  # CM is happy
            else:
                err_msg = self.err_msg or f"Did not raise: {self.expected_exc_types}"
                raise AssertionError(err_msg)

        if not any(
            issubclass(exc_type, expected_type)
            for expected_type in self.expected_exc_types
        ):
            if self.err_msg is not None:
                raise AssertionError(self.err_msg) from exc_value
            else:
                return False  # will re-raise the original exception

        if self.matches is not None:
            err_msg = self.err_msg or (
                "The error message should contain one of the following "
                "patterns:\n{}\nGot {}".format("\n".join(self.matches), str(exc_value))
            )
            if not any(re.search(match, str(exc_value)) for match in self.matches):
>               raise AssertionError(err_msg) from exc_value
E               AssertionError: `RandomFourierKernelApprox.transform()` does not check for consistency between input number
E               of features with RandomFourierKernelApprox.fit(), via the `n_features_in_` attribute.
E               You might want to use `sklearn.utils.validation.validate_data` instead
E               of `check_array` in `RandomFourierKernelApprox.fit()` and RandomFourierKernelApprox.transform()`. This can be done
E               like the following:
E               from sklearn.utils.validation import validate_data
E               ...
E               class MyEstimator(BaseEstimator):
E                   ...
E                   def fit(self, X, y):
E                       X, y = validate_data(self, X, y, ...)
E                       ...
E                       return self
E                   ...
E                   def transform(self, X):
E                       X = validate_data(self, X, ..., reset=False)
E                       ...
E                   return X

.venv/lib/python3.11/site-packages/sklearn/utils/_testing.py:1114: AssertionError
```

There are also a series of `sklearn` deprecation warnings raised, which possibly points to this being due to some breaking change in `sklearn`?
