### 2025-05-10 decargroup/pykoop#188 @sdahdah
Thanks for pointing this out. `sklearn` has a set of built-in unit tests to make sure custom estimators fit their interface. These tests are quite strict (they match exact error messages) and they tend to change with new releases. I've updated everything to match their new tests, and I've also pinned a range of `sklearn` versions. It now works on my machine for Python 3.13, but the CI will tell us if it works for the other versions... There are some remaining deprecation warnings but they are upstream (in `picos`).
