### 2025-05-10 decargroup/pykoop#188 @sdahdah
Please take a look at the linked PR and let me know if you are happy with the changes.
