### 2025-05-10 b0afae4
Remove redundant link in paper
