+++ paper.md
@@ -40,8 +40,6 @@ controller or observer synthesis for a wide range of systems using
 well-established linear tools. `pykoop`'s documentation, along with examples in
 script and notebook form, can be found at at
 [pykoop.readthedocs.io/en/stable](https://pykoop.readthedocs.io/en/stable/).
-Its source code and issue tracker are available at
-[github.com/decargroup/pykoop](https://github.com/decargroup/pykoop).
 Its releases are also archived on Zenodo [@pykoop].
 
 # Statement of need
