### 2025-05-10 93c4c2a
Use non-parenthesized references
