+++ paper.md
@@ -80,10 +80,10 @@ exogenous inputs are not.
 
 # Scholarly publications using `pykoop`
 
-The Koopman operator regression methods proposed in [@dahdah_system_2022] have
+The Koopman operator regression methods proposed in @dahdah_system_2022 have
 been implemented within `pykoop`, while the methods proposed in
-[@dahdah_2024_closed-loop], [@lortie_2024_forward-backward],
-[@lortie_2024_asymptotically], and [@dahdah_2024_uncertainty] are all based on
+@dahdah_2024_closed-loop, @lortie_2024_forward-backward,
+@lortie_2024_asymptotically, and @dahdah_2024_uncertainty are all based on
 `pykoop`, but are implemented in their own repositories.
 
 # Acknowledgements
