### 2025-05-10 136de59
Remove abbreviations from references
