+++ paper.bib
@@ -4,7 +4,7 @@
               and Thirion, B. and Grisel, O. and Blondel, M. and Prettenhofer, P.
               and Weiss, R. and Dubourg, V. and Vanderplas, J. and Passos, A. and
               Cournapeau, D. and Brucher, M. and Perrot, M. and Duchesnay, E.},
-    journal = {J. Mach. Learn. Res.},
+    journal = {Journal of Machine Learning Research},
     volume = {12},
     pages = {2825--2830},
     year = {2011},
@@ -29,7 +29,7 @@
 @inproceedings{dlkoopman,
     author = {Sourya Dey and Eric William Davis},
     title = {{DLKoopman: A deep learning software package for Koopman theory}},
-    booktitle = {Proc. Mach. Learn. Res.},
+    booktitle = {Proceedings of Machine Learning Research},
     pages = {1467--1479},
     volume = {211},
     publisher = {{PMLR}},
@@ -53,13 +53,13 @@
     author = {Steven Dahdah and James R. Forbes},
     title = {System norm regularization methods for {Koopman} operator
              approximation},
-    journal = {Proc. R. Soc. A},
+    journal = {Proceedings of the Royal Society A},
 }
 
 @article{dahdah_2024_closed-loop,
     title = {Closed-loop {Koopman} operator approximation},
     author = {Dahdah, Steven and Forbes, James Richard},
-    journal = {Mach. Learn.: Sci. Technol.},
+    journal = {Machine Learning: Science and Technology},
     volume = {5},
     number = {2},
     pages = {025038},
@@ -89,9 +89,13 @@
     title = {Asymptotically Stable Data-Driven {Koopman} Operator Approximation
         with Inputs using Total Extended {DMD}},
     author = {Lortie, Louis and Forbes, James Richard},
-    year = {2024},
-    journaltitle = {{\tt arXiv:2408.16846v1[eess.SY]}},
-    doi = {10.48550/arXiv.2408.16846},
+    journal = {Machine Learning: Science and Technology},
+    volume = {6},
+    number = {1},
+    pages = {015003},
+    year = {2025},
+    publisher = {{IOP} Publishing},
+    doi={10.1088/2632-2153/ada33b},
 }
 
 @misc{pykoop,
