### 2025-05-10 decargroup/pykoop#187 @sdahdah
Good points, thanks. I've merged these changes into the `paper` branch and generated a new PDF in the review issue.
