### 2025-05-12 decargroup/pykoop#187 @matt-graham
Thanks for fixing these - updated paper all looks good to me!
