### 2025-05-12 #7947 @matt-graham
@gkthiruvathukal Just to let you know that my review checklist is now complete and @sdahdah has addressed the two minor issues I raised (decargroup/pykoop#188 should be resolved once decargroup/pykoop#189 is merged).
