### 2025-06-17 #7947 @kramea
Hello, I can complete it by this Saturday.

Apologies for the delay.

-Kalai

On Mon, Jun 16, 2025 at 7:21 PM George K. Thiruvathukal <
***@***.***> wrote:

> *gkthiruvathukal* left a comment (openjournals/joss-reviews#7947)
> <https://github.com/openjournals/joss-reviews/issues/7947#issuecomment-2978713756>
>
> @kramea <https://github.com/kramea> Can you provide an update on your
> timeframe to complete this JOSS review?
>
