### 2025-07-10 #7947 @kramea
I completed my checklist. The codebase is pretty strong, and well documented. A few minor comments on the paper:

- It would be great if it had some discussions on how pykoop differs from other Koopman packages on practical use cases
- Does it offer any interoperability with other Koopman packages?
