### 2025-07-10 #7947 @kramea
Thanks, @sdahdah. I was thinking about use cases (not features per se, e.g. design etc.). How would it be different if I have to use pykoop for those vs. others? But again, that's a minor comment compared to others.
