### 2025-07-12 a894e77
Update paper with compatibility note and use cases
