### 2025-10-03 #7947 @gkthiruvathukal
I've gone through the editor checklist and believe this is ready to go. I'm proceeding to recommend acceptance.
